Move get_project_config into testing.config
